Edulearn_AI/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
pymupdf
faiss-cpu
